Move the rollout pipeline into its own file
